🛠️ refactor: 코드 리펙토링
- optionExtra int 로 변경

🔗 Resolves: #142

+++ backend/JiShop/src/main/java/com/jishop/option/dto/GeneralOptionResponse.java
@@ -14,7 +14,7 @@ public record GeneralOptionResponse(
             ProductOption productOption = new ProductOption(
                     (Long) option.get("saleProductId"),
                     (String) option.get("optionValue"),
-                    option.get("optionExtra")
+                    (int) option.get("optionExtra")
             );
             generalOptions.add(productOption);
         }
